+++ app/lib/src/model/app_config.dart
@@ -1,21 +1,23 @@
-import 'package:sizzle_starter/src/model/environment.dart';
+import 'package:sizzle_starter/src/model/app_env.dart';
 
-/// Application configuration
-class ApplicationConfig {
-  /// Creates a new [ApplicationConfig] instance.
-  const ApplicationConfig();
+/// App configuration
+/// 
+/// Configurations are read from environment variables, which are set with --dart-define.
+class AppConfig {
+  /// Creates a new [AppConfig] instance.
+  const AppConfig();
 
   /// The current environment.
-  Environment get environment {
+  AppEnv get environment {
     var env = const String.fromEnvironment('ENVIRONMENT').trim();
 
     if (env.isNotEmpty) {
-      return Environment.from(env);
+      return AppEnv.from(env);
     }
 
     env = const String.fromEnvironment('FLUTTER_APP_FLAVOR').trim();
 
-    return Environment.from(env);
+    return AppEnv.from(env);
   }
 
   /// The Sentry DSN.
@@ -25,11 +27,11 @@ class ApplicationConfig {
   bool get enableSentry => sentryDsn.isNotEmpty;
 }
 
-/// A special version of [ApplicationConfig] that is used in tests.
+/// A special version of [AppConfig] that is used in tests.
 ///
-/// In order to use [ApplicationConfig] in tests, it is needed to
+/// In order to use [AppConfig] in tests, it is needed to
 /// extend this class and provide the dependencies that are needed for the test.
-base class TestConfig implements ApplicationConfig {
+base class TestConfig implements AppConfig {
   const TestConfig();
 
   @override