+++ common/common_error_reporter/lib/common_error_reporter.dart
@@ -1,2 +1,2 @@
 export 'src/error_reporter.dart';
-export 'src/sentry_error_reporter.dart';
+export 'src/noop_error_reporter.dart';