+++ common/common_error_reporter/lib/src/error_reporter.dart
@@ -8,18 +8,6 @@ import 'package:common_logger/common_logger.dart';
 /// Implementations should report errors to a service like Sentry/Crashlytics.
 /// {@endtemplate}
 abstract interface class ErrorReporter {
-  /// Returns `true` if the error reporting service is initialized
-  /// and ready to report errors.
-  ///
-  /// If this returns `false`, the error reporting service should not be used.
-  bool get isInitialized;
-
-  /// Initializes the error reporting service.
-  Future<void> initialize();
-
-  /// Closes the error reporting service.
-  Future<void> close();
-
   /// Capture an exception to be reported.
   ///
   /// The [throwable] is the exception that was thrown.
@@ -39,10 +27,6 @@ final class ErrorReporterLogObserver with LogObserver {
 
   @override
   void onLog(LogMessage logMessage) {
-    // If the error reporter is not initialized, do nothing
-    if (!_errorReporter.isInitialized) return;
-
-    // If the log level is error or higher, report the error
     if (logMessage.level.index >= LogLevel.error.index) {
       _errorReporter.captureException(
         throwable: logMessage.error ?? ReportedMessageException(logMessage.message),